test: spliting events tests
